go: move common folder to lib/go
