+++ lib/go/eth/asset_data.go
