+++ lib/go/eth/asset_data_test.go
