+++ lib/go/eth/generate.go
