+++ lib/go/eth/salt.go
